+++ src/svg/enum_svg_color_keywords.py
