+++ src/svg/svg_file.py
@@ -6,13 +6,13 @@ from __future__ import annotations
 import os
 import xml.etree.ElementTree as ET
 
-import svg_writers as sw
-import svg_element_utils as seu
+import svg.svg_writers as sw
+import svg.svg_element_utils as seu
 import file_handlers as fh
 
 from file_handlers import InvalidAccessModeError
 
-from svg_elements import (
+from svg.svg_elements import (
     SVGElement,
     svg,
     g,
