+++ src/svg/svg_validators.py
@@ -4,7 +4,7 @@ appear until the user tries to open it with inkscape or a browser
 
 import re
 import trigonometry as trig
-from enum_svg_color_keywords import ColorKey
+from svg.enum_svg_color_keywords import ColorKey
 
 float_re = "[+-]?[0-9]*\.[0-9]+"
 integer_re = "[+-]?[0-9]+"
