+++ tests/svg_d_attribute_parsing_test.py
@@ -5,7 +5,7 @@ import unittest
 sys.path.append('src')
 
 #from svg_interface import SVGPath
-import svg_parsers as sp
+import svg.svg_parsers as sp
 
 class TestSVGPath(unittest.TestCase):
